# [flutter_tools] implement build aot in terms of assemble for iOS