## Description

Refactors the iOS profile/release script to delegate to assemble. This allows us to incrementally move xcode_backend.sh from bash to dart.

https://github.com/flutter/flutter/issues/32925