@jonahwilliams 2020-01-21
tested via existing iOS devicelab tests and previously written unit tests for targets/ios.dart