@jmagman 2020-01-21
This caused some errant status output for build ios-framework.
Before:
```
[build_ios_framework_module_test] [STDOUT] stdout: Building frameworks for io.flutter.devicelab.hello in debug mode...
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Populating Flutter.framework...                                  0.2s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Creating Flutter.xcframework...                                  1.7s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Adding placeholder App.framework for debug...                   152ms
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Assembling Flutter resources for App.framework...                6.0s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Creating App.xcframework...                                      0.9s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Building plugins...                                             11.5s
[build_ios_framework_module_test] [STDOUT] stdout:  └─Moving to flutter-frameworks/Debug                               0.0s
[build_ios_framework_module_test] [STDOUT] stdout: Building frameworks for io.flutter.devicelab.hello in profile mode...
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Populating Flutter.framework...                                  0.2s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Creating Flutter.xcframework...                                  0.5s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Building Dart AOT for App.framework...                          18.7s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Assembling Flutter resources for App.framework...                0.1s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Creating App.xcframework...                                      1.6s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Building plugins...                                              6.6s
[build_ios_framework_module_test] [STDOUT] stdout:  └─Moving to flutter-frameworks/Profile                             0.0s
[build_ios_framework_module_test] [STDOUT] stdout: Building frameworks for io.flutter.devicelab.hello in release mode...
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Populating Flutter.framework...                                  0.8s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Creating Flutter.xcframework...                                  2.7s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Building Dart AOT for App.framework...                          79.8s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Assembling Flutter resources for App.framework...                0.1s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Creating App.xcframework...                                      0.7s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Building plugins...                                              4.6s
[build_ios_framework_module_test] [STDOUT] stdout:  └─Moving to flutter-frameworks/Release                             0.0s
```
After:
```
[build_ios_framework_module_test] [STDOUT] stdout: Building frameworks for io.flutter.devicelab.hello in debug mode...
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Populating Flutter.framework...                                  0.1s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Creating Flutter.xcframework...                                  1.3s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Adding placeholder App.framework for debug...                   146ms
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Assembling Flutter resources for App.framework...                5.8s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Creating App.xcframework...                                      0.7s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Building plugins...                                              6.0s
[build_ios_framework_module_test] [STDOUT] stdout:  └─Moving to flutter-frameworks/Debug                               0.0s
[build_ios_framework_module_test] [STDOUT] stdout: Building frameworks for io.flutter.devicelab.hello in profile mode...
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Populating Flutter.framework...                                  0.2s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Creating Flutter.xcframework...                                  0.5s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Building Dart AOT for App.framework...                       
[build_ios_framework_module_test] [STDOUT] stdout: Building App.framework for arm64...
[build_ios_framework_module_test] [STDOUT] stdout: 
[build_ios_framework_module_test] [STDOUT] stdout: Building App.framework for armv7...
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Building Dart AOT for App.framework...                          19.9s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Assembling Flutter resources for App.framework...                0.1s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Creating App.xcframework...                                      1.5s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Building plugins...                                              4.5s
[build_ios_framework_module_test] [STDOUT] stdout:  └─Moving to flutter-frameworks/Profile                             0.0s
[build_ios_framework_module_test] [STDOUT] stdout: Building frameworks for io.flutter.devicelab.hello in release mode...
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Populating Flutter.framework...                                  0.9s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Creating Flutter.xcframework...                                  2.0s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Building Dart AOT for App.framework...                       
[build_ios_framework_module_test] [STDOUT] stdout: Building App.framework for arm64...
[build_ios_framework_module_test] [STDOUT] stdout: 
[build_ios_framework_module_test] [STDOUT] stdout: Building App.framework for armv7...
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Building Dart AOT for App.framework...                          86.4s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Assembling Flutter resources for App.framework...                0.0s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Creating App.xcframework...                                      1.1s
[build_ios_framework_module_test] [STDOUT] stdout:  ├─Building plugins...                                              4.9s
[build_ios_framework_module_test] [STDOUT] stdout:  └─Moving to flutter-frameworks/Release                             0.0s
```
It's passing in `quiet = true`
```dart
  await aotBuilder.build(
    platform: TargetPlatform.ios,
    outputPath: iPhoneBuildOutput.path,
    buildMode: mode,
    // Relative paths show noise in the compiler https://github.com/dart-lang/sdk/issues/37978.
    mainDartFile: globals.fs.path.absolute(targetFile),
    quiet: true,
    bitcode: true,
    reportTimings: false,
    iosBuildArchs: <DarwinArch>[DarwinArch.armv7, DarwinArch.arm64],
    dartDefines: dartDefines,
  );
```