@jonahwilliams 2020-01-21
Good catch, will take a look and fix it.