@jonahwilliams 2020-01-22
The snapshotter calls now correctly pass quiet.